perm check for renv and rvar

+++ dev-tools-backend/internal/api/renv/renv.go
@@ -6,8 +6,10 @@ import (
 	"dev-tools-backend/internal/api"
 	"dev-tools-backend/internal/api/middleware/mwauth"
 	"dev-tools-backend/internal/api/middleware/mwcompress"
+	"dev-tools-backend/internal/api/rworkspace"
 	"dev-tools-backend/pkg/idwrap"
 	"dev-tools-backend/pkg/service/senv"
+	"dev-tools-backend/pkg/service/suser"
 	"dev-tools-backend/pkg/translate/tenv"
 	"dev-tools-backend/pkg/translate/tgeneric"
 	environmentv1 "dev-tools-services/gen/environment/v1"
@@ -20,6 +22,7 @@ type EnvRPC struct {
 	DB *sql.DB
 
 	es senv.EnvService
+	su suser.UserService
 }
 
 func CreateService(ctx context.Context, db *sql.DB, secret []byte) (*api.Service, error) {
@@ -44,21 +47,33 @@ func CreateService(ctx context.Context, db *sql.DB, secret []byte) (*api.Service
 	return &api.Service{Path: path, Handler: handler}, nil
 }
 
-// TODO: add perm checks
 func (e *EnvRPC) CreateEnvironment(ctx context.Context, req *connect.Request[environmentv1.CreateEnvironmentRequest]) (*connect.Response[environmentv1.CreateEnvironmentResponse], error) {
 	envReq := tenv.DeseralizeRPCToModelWithID(idwrap.NewNow(), req.Msg.GetEnvironment())
-	err := e.es.Create(ctx, envReq)
+	ok, err := rworkspace.CheckOwnerWorkspace(ctx, e.su, envReq.WorkspaceID)
+	if err != nil {
+		return nil, connect.NewError(connect.CodeInternal, err)
+	}
+	if !ok {
+		return nil, connect.NewError(connect.CodePermissionDenied, nil)
+	}
+	err = e.es.Create(ctx, envReq)
 	if err != nil {
 		return nil, connect.NewError(connect.CodeInternal, err)
 	}
 	return connect.NewResponse(&environmentv1.CreateEnvironmentResponse{Id: envReq.ID.String()}), nil
 }
 
-// TODO: add perm checks
 func (e *EnvRPC) GetEnvironment(ctx context.Context, req *connect.Request[environmentv1.GetEnvironmentRequest]) (*connect.Response[environmentv1.GetEnvironmentResponse], error) {
 	id, err := idwrap.NewWithParse(req.Msg.Id)
 	if err != nil {
 		return nil, connect.NewError(connect.CodeInvalidArgument, err)
+	}
+	ok, err := CheckOwnerEnv(ctx, e.su, e.es, id)
+	if err != nil {
+		return nil, connect.NewError(connect.CodeInternal, err)
+	}
+	if !ok {
+		return nil, connect.NewError(connect.CodePermissionDenied, nil)
 	}
 	env, err := e.es.Get(ctx, id)
 	if err != nil {
@@ -68,11 +83,17 @@ func (e *EnvRPC) GetEnvironment(ctx context.Context, req *connect.Request[enviro
 	return connect.NewResponse(&environmentv1.GetEnvironmentResponse{Environment: resp}), nil
 }
 
-// TODO: add perm checks
 func (e *EnvRPC) GetEnvironments(ctx context.Context, req *connect.Request[environmentv1.GetEnvironmentsRequest]) (*connect.Response[environmentv1.GetEnvironmentsResponse], error) {
 	workspaceID, err := idwrap.NewWithParse(req.Msg.WorkspaceId)
 	if err != nil {
 		return nil, connect.NewError(connect.CodeInvalidArgument, err)
+	}
+	ok, err := rworkspace.CheckOwnerWorkspace(ctx, e.su, workspaceID)
+	if err != nil {
+		return nil, connect.NewError(connect.CodeInternal, err)
+	}
+	if !ok {
+		return nil, connect.NewError(connect.CodePermissionDenied, nil)
 	}
 
 	envs, err := e.es.GetByWorkspace(ctx, workspaceID)
@@ -84,11 +105,18 @@ func (e *EnvRPC) GetEnvironments(ctx context.Context, req *connect.Request[envir
 	return connect.NewResponse(&environmentv1.GetEnvironmentsResponse{Environments: resp}), nil
 }
 
-// TODO: add perm checks
 func (e *EnvRPC) UpdateEnvironment(ctx context.Context, req *connect.Request[environmentv1.UpdateEnvironmentRequest]) (*connect.Response[environmentv1.UpdateEnvironmentResponse], error) {
 	envReq, err := tenv.DeserializeRPCToModel(req.Msg.GetEnvironment())
 	if err != nil {
 		return nil, connect.NewError(connect.CodeInvalidArgument, err)
+	}
+
+	ok, err := CheckOwnerEnv(ctx, e.su, e.es, envReq.ID)
+	if err != nil {
+		return nil, connect.NewError(connect.CodeInternal, err)
+	}
+	if !ok {
+		return nil, connect.NewError(connect.CodePermissionDenied, nil)
 	}
 
 	err = e.es.Update(ctx, &envReq)
@@ -99,15 +127,35 @@ func (e *EnvRPC) UpdateEnvironment(ctx context.Context, req *connect.Request[env
 	return connect.NewResponse(&environmentv1.UpdateEnvironmentResponse{}), nil
 }
 
-// TODO: add perm checks
 func (e *EnvRPC) DeleteEnvironment(ctx context.Context, req *connect.Request[environmentv1.DeleteEnvironmentRequest]) (*connect.Response[environmentv1.DeleteEnvironmentResponse], error) {
 	id, err := idwrap.NewWithParse(req.Msg.Id)
 	if err != nil {
 		return nil, connect.NewError(connect.CodeInvalidArgument, err)
+	}
+	ok, err := CheckOwnerEnv(ctx, e.su, e.es, id)
+	if err != nil {
+		return nil, connect.NewError(connect.CodeInternal, err)
+	}
+	if !ok {
+		return nil, connect.NewError(connect.CodePermissionDenied, nil)
 	}
 	err = e.es.Delete(ctx, id)
 	if err != nil {
 		return nil, connect.NewError(connect.CodeInternal, err)
 	}
 	return connect.NewResponse(&environmentv1.DeleteEnvironmentResponse{}), nil
+}
+
+func CheckOwnerEnv(ctx context.Context, su suser.UserService, es senv.EnvService, envid idwrap.IDWrap) (bool, error) {
+	env, err := es.Get(ctx, envid)
+	if err != nil {
+		return false, err
+	}
+
+	userID, err := mwauth.GetContextUserID(ctx)
+	if err != nil {
+		return false, err
+	}
+
+	return su.CheckUserBelongsToWorkspace(ctx, userID, env.WorkspaceID)
 }
